+++ NEWS.md
@@ -8,6 +8,7 @@
 * 21 palettes have been added from the peRReo package. (#122)
 * 30 palettes have been added from the NatParksPalettes package. (#123)
 * 24 palettes have been added from the impressionist.colors package. (#125)
+* 7 palettes have been added from the severance package. (#127)
 
 # paletteer 1.4.1